% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADeval.r
\name{ADeval}
\alias{ADeval}
\title{ADeval}
\usage{
ADeval(simdat, zcutoff, finalm = c("z", "ci"), alpha = 0.025)
}
\arguments{
\item{simdat}{matrix or data.frame of multiple runs of ADsim}

\item{zcutoff}{a matrix (or data frame) with the cut-offs on the z-statstics at interim.
One row per interim analysis and a column for stop for futilty and efficacy (with NA if not planned)}

\item{finalm}{Measure used for the final analysis, either z (z-statistics) or ci (confidence interval).}

\item{alpha}{Alpha level used for the final analysis if finalm=="z", otherwise the ci given in simdat is used.}
}
\value{
A list with the evaluation for each repetition (\strong{rawdata}) and the summarized operating characteristics (\strong{opchar}).
\strong{opchar} includes the average estimates in each group, effect (point estimate, \strong{pe}) and (1-alpha) confidence limits (\strong{lci}, \strong{uci}), and
\itemize{
\item \strong{psig}: the probability for a significant trial (the power or type I error, depending on the assumptions).
\item \strong{pstop}: the probability for any stop and a stop for futility or efficacy (\strong{pstop_fut}, \strong{pstop_eff}).
\item the average, minimum and maximum sample size (\strong{avn}, \strong{minn}, \strong{maxn}).
\item \strong{bias}: the average of estimate minus true value.
\item \strong{sd}: the standard deviation of the point estimate.
\item \strong{mse}: the mean squared error, the average of the squared difference between estimate and true value.
\item \strong{coverage}: the proportion of confidence intervals including the true value.
}
}
\description{
Function to evaluate operating characteristics of trials simulated via ADsim
}
\examples{
#Simulate 1000 trials with one interim analysis
simdat <- lapply(1:1000, function(i)
 ADsim(n01 = c(100,100), p01 = c(0.4, 0.2), nia = 1))
 simdat <- do.call(rbind, simdat)

# Define stop for futility stoppin (z<0) and efficacy stopping (z>2)
zcutoff<-matrix(c(0,2),1,2)
sr<-ADeval(simdat = simdat, zcutoff = zcutoff)
head(sr$rawdata)
sr$opchar


}
