% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADsimfun.r
\name{ADsimfun}
\alias{ADsimfun}
\title{ADsimfun}
\usage{
ADsimfun(
  n01,
  p01,
  nia,
  tia = NULL,
  alpha = 0.025,
  direct = c("lower", "higher"),
  effm = c("rd", "rr"),
  cimethod = c("Wald", "score")
)
}
\arguments{
\item{n01}{Total sample size in each arm, numeric vector of length 2}

\item{p01}{Proportion in each arm, numeric vector of length 2}

\item{nia}{Number of interim analysis, not including the final analysis}

\item{tia}{Optional vector with the time point(s) of the interim analyses, as fraction of the total (between 0 and 1)}

\item{alpha}{Alpha level for final analysis}

\item{direct}{Direction of effect, "lower" (lower is better) or higher ("higher is better")}

\item{effm}{Effect measure, "rd" (risk differnce) or "rr" (risk ratio)}

\item{cimethod}{Confidence interval used for analysis, "Wald" or "score".}
}
\value{
Numeric vector with true proportions in both arms and point estimate (true_p0, true_p1, true_pe),
and for each interim stage (prefix iax_) and the final stage (prefix fa_),
the simulated number of successes, sample size and observed probabilities for both arms (suffix x0, x1, n0, n1, p0, p1),
point estimate (suffic pe), confidence interval (suffic lci and uci) and z-statistic (z).
}
\description{
Simulates trials with a binary endpoint and one or several interim analysis,
}
\examples{
set.seed(1)
ADsimfun(n01 = c(100,100), p01 = c(0.4,0.2), nia = 1, alpha = 0.025)
set.seed(1)
ADsimfun(n01 = c(100,100), p01 = c(0.4,0.2), nia = 1, tia = 0.4, alpha = 0.025)
set.seed(1)
ADsimfun(n01 = c(100,100), p01 = c(0.4,0.2), nia = 1, alpha=0.025, effm="rr")
set.seed(1)
ADsimfun(n01 = c(100,100), p01 = c(0.4,0.2), nia = 2, alpha=0.025,
  direct="higher", effm="rd")
}
