% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.r
\name{launch_ADSim_app}
\alias{launch_ADSim_app}
\title{AdaptiveDesignSim shiny app}
\usage{
launch_ADSim_app()
}
\description{
The shiny app to AdaptiveDesignSim provides an intuitive way to
\enumerate{
\item check the results of individual simulations (via forest plots) and
\item of operating characteristics over a range of conditions.
}
}
\examples{
# launch the app
\dontrun{
launch_ADSim_app()
}

}
