% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADevalfun.r
\name{ADevalfun}
\alias{ADevalfun}
\title{ADevalfun}
\usage{
ADevalfun(simdat, zcutoff)
}
\arguments{
\item{simdat}{matrix or data.frame of multiple runs of ADsimfun}

\item{zcutoff}{a matrix (or data frame) with the cut-offs on the z-statstics at interim.
One row per interim analysis and a column for stop for futilty and efficacy (with NA if not planned)}
}
\value{
A list with the evaluation for each repetition (\strong{rawdata}) and the summarized operating characteristics (\strong{opchar}).
\strong{opchar} includes the average estimates in each group, effect (point estimate, \strong{pe}) and (1-alpha) confidence limits (\strong{lci}, \strong{uci}), and
\itemize{
\item \strong{psig}: the probability for a significant trial (the power or type I error, depending on the assumptions).
\item \strong{pstop}: the probability for any stop and a stop for futility or efficacy (\strong{pstop_fut}, \strong{pstop_eff}).
\item the average, minimum and maximum sample size (\strong{avn}, \strong{minn}, \strong{maxn}).
\item \strong{bias}: the average of estimate minus true value.
\item \strong{sd}: the standard deviation of the point estimate.
\item \strong{mse}: the mean squared error, the average of the squared difference between estimate and true value.
\item \strong{coverage}: the proportion of confidence intervals including the true value.
}
}
\description{
Function to evaluate operating characteristics of trials simulated via ADsimfun
}
\examples{
#Simulate 1000 trials with one interim analysis
simdat <- lapply(1:1000, function(i)
 ADsimfun(n01 = c(100,100), p01 = c(0.4, 0.2), nia = 1, alpha = 0.025, effm="rd", direct="lower"))
 simdat <- do.call(rbind, simdat)

# Define stop for futility stoppin (z<0) and efficacy stopping (z>2)
zcutoff<-matrix(c(0,2),1,2)
sr<-ADevalfun(simdat = simdat, zcutoff = zcutoff)
head(sr$rawdata)
sr$opchar


}
