% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADsim.r
\name{ADsim}
\alias{ADsim}
\title{ADsim}
\usage{
ADsim(
  n01,
  p01,
  nia,
  tia = NULL,
  estfun = "wald_rd",
  cilevel = 0.95,
  truefun = NULL,
  direct = c("lower", "higher")
)
}
\arguments{
\item{n01}{Total sample size in each arm, numeric vector of length 2}

\item{p01}{Proportion in each arm, numeric vector of length 2}

\item{nia}{Number of interim analysis, not including the final analysis}

\item{tia}{Optional vector with the time point(s) of the interim analyses, as fraction of the total (between 0 and 1)}

\item{estfun}{Function to estimate to determine point estimate, CIs and test.
Should take arguments xs (number of success in both groups), ns (sample size in both groups), and cilevel,
and return a named vector with pe, lci, uci and z.
Four defaults functions can be entered as string: wald_rd, wald_rr, score_rd, score_rr.}

\item{cilevel}{Level of the two-sided confience interval for the default functions.}

\item{truefun}{Function to calculate the true point estimate from p01 (ignored if the default estfun are used)}

\item{direct}{Direction of the effect, "lower" or "higher" values are better.}
}
\value{
Numeric vector with true proportions in both arms and point estimate (true_p0, true_p1, true_pe),
and for each interim stage (prefix iax_) and the final stage (prefix fa_),
the simulated number of successes, sample size and observed probabilities for both arms (suffix x0, x1, n0, n1, p0, p1),
point estimate (suffic pe), confidence interval (suffic lci and uci) and z-statistic (z).
}
\description{
Simulates trials with a binary endpoint and one or several interim analysis,
}
\examples{
set.seed(1)
ADsim(n01 = c(100,100), p01 = c(0.4,0.2), nia = 1)

#Earlier interim analysis
set.seed(1)
ADsim(n01 = c(100,100), p01 = c(0.4,0.2), nia = 1, tia = 0.4)

#Difference effect estimation
set.seed(1)
ADsim(n01 = c(100,100), p01 = c(0.4,0.2), nia = 1, estfun = "wald_rr")

#Assume higher would be better
set.seed(1)
ADsim(n01 = c(100,100), p01 = c(0.4,0.2), nia = 2, direct="higher")
efun<-function(xs,ns,cilevel) {
pe<-diff(xs/ns)
se<-sqrt(sum(xs/ns*(1-xs/ns)/ns))
lci<-pe - qnorm(1-(1-cilevel)/2)*se
uci<-pe + qnorm(1-(1-cilevel)/2)*se
z<-(-pe)/se
res<-c(pe,lci,uci,z)
names(res)<-c("pe","lci","uci","z")
return(res)
}
set.seed(1)
ADsim(n01 = c(100,100), p01 = c(0.4,0.2), nia = 1, estfun = efun, truefun = function(x) diff(x))

}
